feat: add header and main landing page
