+++ suspect-game/package.json
@@ -26,6 +26,7 @@
     "react-material-ui-carousel": "^3.4.2",
     "react-modal": "^3.16.1",
     "react-modal-video": "^2.0.1",
+    "react-type-animation": "^3.1.0",
     "sass": "^1.63.6",
     "swiper": "^9.3.2",
     "tailwindcss": "3.3.1",
